requirements.txt
flask
sqlalchemy
flask-sqlalchemy
flask-login 
cryptography
PyJWT
python-dateutil
requests
apscheduler
keyring
Faker
secretstorage; sys_platform == 'linux'